requirements.txt
flask==2.3.3
unstructured==0.10.30
pytesseract==0.3.10
pdf2image==1.16.3
langchain>=0.0.350
langchain-text-splitters>=0.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
python-dotenv==1.0.0 